chore: group controllers into packages by tier
