+++ railgun/controllers/inputs/utils.go
@@ -1,4 +1,4 @@
-package controllers
+package inputs
 
 import (
 	"context"
